Stack-Only Data: Copy
Rust has a special annotation called the Copy trait that we can place on types that are stored on the stack, in this example integer.
If a type implements the Copy trait, variables that use it do not move, but rather are trivially copied, making them still valid after assignment to another variable.