Slice type
A function that takes a string of words separated by spaces and return the index of the end of the word, indicated by a space.
We could use that value 5 (word) with the variable s to try to extract the first word out, but this would be a bug because the contents of s have changed since we saved 5 in word.