Struct update
Case: create a new instance of a struct that includes most of the values from another instance, but changes some.
The struct update syntax uses = like an assignment; this is because it moves the data.
We can no longer use user1 after creating user5 because the String in the username field of user1 was moved into user5.