Print the value of struct instance (for debugging)
Use dbg! macro.
dbg! takes ownership of an expression (as opposed to println! that takes a reference), prints the file and line number of where that dbg! macro call occurs in your code along with the resulting value of that expression, and returns ownership of the value.
Calling the dbg! macro prints to the standard error console stream (stderr), as opposed to println! which prints to the standard output console stream (stdout).