Method syntax
Methods are defined within the context of a struct (or an enum or a trait object, respectively), and their first parameter is always self, which represents the instance of the struct the method is being called on.
Within an impl block, the type Self is an alias for the type that the impl block is for.
The main reason for using methods instead of functions is for organization. We’ve put all the things we can do with an instance of a type in one impl block rather than making future users of our code. search for capabilities of Rectangle in various places in the library we provide.