Paths for Referring to an Item in the Module Tree
Using the crate name to start from the crate root is like using / to start from the filesystem root in your shell.
Choosing whether to use a relative or absolute path is a decision you’ll make based on your project, and depends on whether you’re more likely to move item definition code separately from or together with the code that uses the item.
Our preference in general is to specify absolute paths because it’s more likely we’ll want to move code definitions and item calls independently of each other.
- If we move the front_of_house module and the eat_at_restaurant function into a module named customer_experience, we’d need to update the absolute path to add_to_waitlist, but the relative path would still be valid.
- If we moved the eat_at_restaurant function separately into a module named dining, the absolute path to the add_to_waitlist call would stay the same, but the relative path would need to be updated.