Module - Exposing Paths with the pub Keyword
Items in a parent module can't use the private items inside child modules (by default), but items in child modules can use the items in their ancestor modules. This is because child modules wrap and hide their implementation details, but the child modules can see the context in which they’re defined.