Exception in idiomatic path
The exception to this idiom is if we’re bringing two items with the same name into scope with use statements.
Using the parent modules distinguishes the two Result types. If instead we specified use std::fmt::Result and use std::io::Result, we’d have two Result types in the same scope and Rust wouldn’t know which one we meant when we used Result.