Accessing vector
There are two ways to reference a value stored in a vector:
- Via indexing (fail fast)
- Using the get method (fail safe)
Why Rust give us 2 methods? We have choices how the program behaves when you try to use an index value outside the range of existing elements.
The different is using get the Rust won't panic, so, we have a choice how to handle None.