String representation
A String is a wrapper over a Vec<u8>.
For last hello string it took 2 bytes of storage to store each Unicode scalar value.
3 -> 208. answer should have 208 value, but it's not a valid character on its own.

So, why string in Rust can't be access via index is to avoid returning an unexpected value and causing bugs that might not be discovered immediately, Rust doesn’t compile this code at all and prevents misunderstandings early in the development process.